4chat/services/search-service/requirements.txt
fastapi
uvicorn[standard]
elasticsearch>=8.0.0,<9.0.0
python-jose[cryptography]
pydantic
python-dotenv
httpx
aiokafka
nacos-sdk-python
celery[redis]>=5.4.0
